新增了README.md 删除了不必要的类
