--- src/main/java/com/deystar/Const/ApplicationConst.java
@@ -1,12 +1,0 @@
-package com.deystar.Const;
-
-/**
- * @Author YeungLuhyun
- **/
-public class ApplicationConst {
-
-    public static final Integer MAX_THREAD_NUM = 10;
-
-    public static final Integer THREAD_NUM = 4;
-
-}
