+++ src/main/java/com/deystar/Const/Suffix.java
@@ -4,6 +4,7 @@ import org.omg.CORBA.PUBLIC_MEMBER;
 
 /**
  * @author Ming Yeung Luhyun (杨名 字 露煊)
+ *  package `s suffix
  */
 public enum Suffix {
