--- src/main/java/com/deystar/Zip/SevenZip/RunZipTask/ZipService.java
@@ -1,10 +1,0 @@
-package com.deystar.Zip.SevenZip.RunZipTask;
-
-/**
- * @author Ming Yeung Luhyun (杨名 字 露煊)
- */
-public interface ZipService extends Runnable{
-
-
-    void run();
-}
